+++ test/function/arithmetic/divide.test.js
@@ -78,15 +78,15 @@ describe('divide', function() {
   }); 
 
   it('should divide 1 over a matrix (matrix inverse)', function() {
-    assert.deepEqual(math.format(divide(1, [
+    approx.deepEqual(divide(1, [
       [ 1, 4,  7],
       [ 3, 0,  5],
       [-1, 9, 11]
-    ])), math.format([
+    ]), [
       [ 5.625, -2.375, -2.5],
       [ 4.75,  -2.25,  -2],
       [-3.375,  1.625,  1.5]
-    ]));
+    ]);
   });
 
   it('should perform matrix division', function() {